Conflicts:
	src/main/java/alkemy/challenge/Challenge/Alkemy/controller/ContactController.java
	src/main/java/alkemy/challenge/Challenge/Alkemy/service/ContactService.java

+++ src/main/java/alkemy/challenge/Challenge/Alkemy/service/OrganizationService.java
@@ -12,12 +12,12 @@ public class OrganizationService {
 
     private final OrganizationRepository organizationRepository;
 
-    public Organization bringOrganization(int id) {
+    public Organization bringOrganization(Long id) {
         return organizationRepository.getById((long) id);
     }
 
     public void update(OrganizationDto organizationDto) {
-        Organization organization = bringOrganization(1);
+        Organization organization = bringOrganization(1L);
         organization.setName(organizationDto.getName());
         organization.setImage(organizationDto.getImage());
         if (organizationDto.getAdress() != null) {
